refactor(ui): extract shared QueryCodeEditor component
